start making datastore an interface
