+++ src/me/ryanhamshire/GriefPrevention/events/config/ConfigManager.java
@@ -1,0 +1,10 @@
+package me.ryanhamshire.GriefPrevention.events.config;
+
+/**
+ * Created on 5/27/2017.
+ *
+ * @author RoboMWM
+ */
+public class ConfigManager
+{
+}
